+++ src/main/java/team/hanaro/hanamate/entities/Images.java
@@ -1,4 +1,4 @@
-package team.hanaro.hanamate.domain;
+package team.hanaro.hanamate.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,20 +6,24 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import javax.persistence.*;
+import java.sql.Blob;
 import java.sql.Timestamp;
 
 @Entity
-@Table(name = "allowance_history")
+@Table(name = "images")
 @Getter
 @ToString
 @AllArgsConstructor
 @NoArgsConstructor
-public class AllowanceHistory {
+public class Images {
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long historyId;
-    private Long allowanceId;
-    private Integer allowanceAmount;
-    private Timestamp transactionDate;
-    private Boolean success;
+    private Long imageId;
+    private String fileName;
+    private Integer fileSize;
+    private String fileType;
+    @Lob
+    private Blob fileData;
+    private Timestamp createdAt;
 }
