+++ src/main/java/team/hanaro/hanamate/entities/SavingsHistory.java
@@ -1,4 +1,4 @@
-package team.hanaro.hanamate.domain;
+package team.hanaro.hanamate.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,17 +9,19 @@ import javax.persistence.*;
 import java.sql.Timestamp;
 
 @Entity
-@Table(name = "loan_history")
+@Table(name = "savings_history")
 @Getter
 @ToString
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoanHistory {
+public class SavingsHistory {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long historyId;
-    private Long allowanceId;
-    private Integer allowanceAmount;
+    private Long savingsId;
+    private Integer sequence;
+    private Integer depositAmount;
     private Timestamp transactionDate;
+    private Integer balance;
     private Boolean success;
 }
